benerin work request + work order

+++ database/migrations/2019_01_07_102119_create_trx_work_order_detail_table.php
@@ -22,7 +22,7 @@ class CreateTrxWorkOrderDetailTable extends Migration
             $table->unsignedInteger('material_id')->nullable();
             $table->unsignedInteger('resource_id')->nullable();
             $table->unsignedInteger('wbs_id')->nullable();
-            $table->bigInteger('total_price')->nullable();
+            $table->double('total_price')->nullable();
             $table->timestamps();
 
             $table->foreign('material_id')->references('id')->on('mst_material');
